Android test262: Added custom log because the adb logcat was not large enough to hold everything due to gc noise.
Code is a mess because of problems actually writing the log and trying to extract it. My confusion of the classes plus Android lying to me about where the storage directory is led to a lot of clutter.

Started adding memory analysis code. Should have committed before adding that.

Internet access is added to the manifest. It is not needed right now, but I’m investigating the possibility of sending the log through the network.