Android test262: Mac client can now receive and display echoed log messages from the device as it is running via a socket connection.
The system reuses the initial port communication for the http server to denote it wants a log stream event instead of an download. From there, the socket is kept open to echo messages through the socket back to the Mac client.

The Mac client opens a new window for each device and displays the log messages in a NSTextView.

Also added Alert dialogs for resolve/connection errors.