Android test262: Mac client fixes/performance enhancements for log streaming.
Thanks to the mailing list, scrollPoint is much faster, and fast enough for our purposes for autoscrolling. The method has been reimplemented (with some fussing for figuring out when autoscroll is on, and where to set the position).

Also figured out the exception error was due to strings passing funny or null characters in the string which caused the Cocoa NSString creation to fail. Defensive checks are now in place.

Also disabled word-wrapping for performance. (Resizing the window with 800k lines would stall the program for awhile.)

Known bug: Around 800,000 lines, the text view stops updating itself with the new appended text (10.9.0). Cmd-A, Cmd-C, and pasting into another app actually gets all the contents including the missing text, so the text is there, and it looks like it is an OS X display bug.