test262: Added 3rd party Logger (playcontrol.net/opensource/Logger) for test262 purposes (MIT). Logging system now writes to hardcoded test262_runlog.txt in Documents directory and echoes to NSLog for Mac and iOS.
The NSLog echoing is intentionally done to deal with the problem that Apple keeps restricting what gets logged with every release. Circa iOS 6, stdout/stderr were thrown away when not launched directly through Xcode. http://www.coronalabs.com/blog/2012/09/19/faq-wednesday-iphone-5-and-ios-6/. Thus NSLog is a defensive option.

As I started writing the logging system, I realized I was reinventing most of what I already wrote in Logger. Logger has also been evolving a set of useful extensions to redirect output which I used to redirect to NSLog, and will also use to redirect to sockets. It is also very friendly with printf-style semantics which work great for NSLog, Obj-C, and C, and Logger already has Obj-C extensions. Thus, to save me a bunch of time, I dropped Logger in here. The dependency is actually quite small. It is basically Logger.c + TimeStamp.c + LoggerObjC.m (for the Obj-C extensions. Logger was designed to be easy to port to esoteric embedded platforms so its implementation is kept fairly small and simple. This could be brought back to Android test262 since the implementation is pure C, so I placed it in the common directory.