CMake: Added -fvisibility-inlines-hidden and new option EXPORT_ONLY_PUBLIC_SYMBOLS which invokes -fvisibility=hidden on gcc/clang platforms.
The old code invokes -fvisibility=hidden only for static libraries which I believe is wrong because I think it essentially becomes a no-op. The feature is intended for dynamic libraries. The change log suggests this was done to avoid problems. My suspicion is that the problem is that jsc and testapi use private JSCore APIs so this broke the build. Applying that only to static libraries had no effect so it avoided breaking the build.

Symbol visibility is important for multiple reasons including code size and reducing potential symbol clashes. I’ve added a new option, EXPORT_ONLY_PUBLIC_SYMBOLS which will enable this behavior and avoid building jsc. Perhaps a better fix is to actually make the private API in use public (and fixes the bad code smell), or build both a static and dynamic library of JSCore and make jsc link against the static version.

I also enabled -fvisibility-inlines-hidden. This is generally a good idea and the change log says this is enabled in both the Mac and configure/Makefile build systems. So we should enable it too.

Note: No changes were done for the Windows builds.